requirements.txt
asyncpixel
discord.py
python-dotenv
quart
pymongo
dnspython
requests
discord-timers
PyNaCl
hypercorn
youtube-dl
discord-pretty-help